Update MyPanel.java
Added a for loop that will be used for planting the 10 mines in a random spot every time. 
This loop works but needs to be worked with to function the way it should.
Trying to implement that when a block with a mine is clicked, the block will become black.
Added 4 public void/int statements that have no use yet.